requirements.txt
Flask==3.0.3
pymongo
dotenv
flask-cors
python-dateutil